벨트와 연동 성공 및 이와 관련한 스크립트들 수정. 낚시 결과 UI 수정, 낚시 시작 화면에서 모드 아이콘 변경.

+++ Hoho/Assets/Scripts/아이모드/낚시게임/낚시진행/FishGenerator.cs
@@ -255,7 +255,7 @@ public class FishGenerator : MonoBehaviour
         {
             //500->0, 850 ->1
 
-            hookPos = (Characteristic.value - BreathingTest.breathingMin) / (BreathingTest.breathingMax - BreathingTest.breathingMin);
+            hookPos = (float) (Characteristic.value - BreathingTest.breathingMin) /(float) (BreathingTest.breathingMax - BreathingTest.breathingMin);
         }
         GameObject.Find("³¬½Ë¹Ù´Ã").GetComponent<HookController>().setPosition(hookPos);
